Struct hash_ring::HashRing

pub struct HashRing<T> {
    // some fields omitted
}

HashRing

Methods

impl<T: ToString + Clone> HashRing<T>

fn new(nodes: Vec<T>, replicas: isize) -> HashRing<T>

Creates a new hash ring with the specified nodes. Replicas is the number of virtual nodes each node has to make a better distribution.

fn add_node(&mut self, node: &T)

Adds a node to the hash ring

fn remove_node(&mut self, node: &T)

Deletes a node from the hash ring

fn get_node(&mut self, key: String) -> Option<&T>

Gets the node a specific key belongs to